+++ NEWS.md
@@ -1,5 +1,7 @@
 # worldbank (development version)
 
+- Support month and quarter date format in `wb_country_indicator()`
+
 # worldbank 0.1.0
 
-* Initial CRAN submission.
+- Initial CRAN submission.